Fix up files and instance logic in SF
