+++ src/Shared/Microsoft.DotNet.Internal.Health/AzureTableHealthReportProvider.cs
@@ -211,24 +211,4 @@ namespace Microsoft.DotNet.Internal.Health
             _client?.Dispose();
         }
     }
-
-    public class HealthReport
-    {
-        public HealthReport(string serviceName, string instance, string subStatusName, HealthStatus health, string message, DateTimeOffset asOf)
-        {
-            Service= serviceName;
-            Instance = instance;
-            SubStatus = subStatusName;
-            Health = health;
-            Message = message;
-            AsOf = asOf;
-        }
-
-        public string Service { get; }
-        public string Instance { get; }
-        public string SubStatus{ get; }
-        public HealthStatus Health { get; }
-        public string Message { get; }
-        public DateTimeOffset AsOf { get; }
-    }
 }
